pyproject.toml
[tool.poetry]
name = "azure-rag-examples"
version = "0.1.0"
description = ""
authors = ["Marcin Dąbrowski"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
azure-storage-blob = "^12.24.1"
azure-functions = "^1.21.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
